tk 5

+++ source/progressdialog.cpp
@@ -132,7 +132,7 @@ void ProgressThread::reportReady()
 {
     if (mainThreadId != QThread::currentThreadId())
         return;
-    QMessageBox::warning(nullptr, "Ready", QString("Just Ready. %1 vs %2").arg(mainThreadId).arg(subThreadId));
+    QMessageBox::warning(nullptr, "Ready", QString("Just Ready. %1 vs %2").arg((int)mainThreadId).arg((int)subThreadId));
     emit this->taskReady();
 }
 
@@ -252,7 +252,7 @@ void ProgressDialog::start(PROGRESS_DIALOG_STATE mode, const QString &label, int
 
 void ProgressDialog::done(bool forceOpen)
 {
-    QMessageBox::warning(nullptr, "Done", QString("Really. %1 vs %2").arg(mainThreadId).arg(subThreadId));
+    QMessageBox::warning(nullptr, "Done", QString("Really. %1 vs %2").arg((int)mainThreadId).arg((int)subThreadId));
 
     theDialog->setWindowTitle(" ");
     theDialog->ui->progressLabel->setVisible(false);
